+++ contracts/src/arbitration/KlerosCoreBase.sol
@@ -1091,8 +1091,12 @@ abstract contract KlerosCoreBase is IArbitratorV2, Initializable, UUPSProxiable 
         courtJump = true;
 
         if (!courts[newCourtID].supportedDisputeKits[newDisputeKitID]) {
-            // Switch to classic dispute kit if parent court doesn't support the current one.
-            newDisputeKitID = DISPUTE_KIT_CLASSIC;
+            // The current Dispute Kit is not compatible with the new court, jump to another Dispute Kit.
+            newDisputeKitID = disputeKits[_round.disputeKitID].getJumpDisputeKitID();
+            if (newDisputeKitID == NULL_DISPUTE_KIT || !courts[newCourtID].supportedDisputeKits[newDisputeKitID]) {
+                // The new Dispute Kit is not defined or still not compatible, fall back to `DisputeKitClassic` which is always supported.
+                newDisputeKitID = DISPUTE_KIT_CLASSIC;
+            }
             disputeKitJump = true;
         }
     }